#8 ajout des composant de DeckChoice
