+++ src/components/CardList.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function CardList ({ heroes }) {
+    return (
+        <div>
+          
+        </div>
+    )
+}
+
+export default CardList;
