+++ src/components/DeckList.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function DeckList ({ heroes }) {
+    return (
+        <div>
+          
+        </div>
+    )
+}
+
+export default DeckList;
