+++ src/components/NavBarDeckChoice.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function NavBarDeckChoice ({ heroes }) {
+    return (
+        <div>
+         
+        </div>
+    )
+}
+
+export default NavBarDeckChoice;
